refactor(e2e): Bring provisioning x509 and symmetric key tests to android and things
