+++ iot-e2e-tests/android/app/build.gradle
@@ -1,13 +1,18 @@
 apply plugin: 'com.android.application'
 
 //***********************************************************************************************//
-def IOTHUB_E2E_X509_PRIVATE_KEY_BASE64 = project.hasProperty('IotHubPrivateKeyBase64') ? '"'+project.property('IotHubPrivateKeyBase64')+'"' : '"Define IotHub Private Key"'
-def IOTHUB_E2E_X509_CERT_BASE64 = project.hasProperty('IotHubPublicCertBase64') ? '"'+project.property('IotHubPublicCertBase64')+'"': '"Define IotHub Public Cert"'
-def IOTHUB_CONNECTION_STRING_ENV_VAR_NAME = project.hasProperty('IotHubConnectionString') ? '"'+project.property('IotHubConnectionString')+'"' : '"Define IotHub Connection"'
-def IOTHUB_E2E_X509_THUMBPRINT = project.hasProperty('IotHubThumbprint') ? '"'+project.property('IotHubThumbprint')+'"': '"Define IotHub Cert Thumbprint"'
-def IOTHUB_CONN_STRING_INVALIDCERT = project.hasProperty('IotHubInvalidCertConnectionString') ? '"'+project.property('IotHubInvalidCertConnectionString')+'"': '"Define IotHub Invalid Cert Connection"'
-def APPCENTER_APP_SECRET = project.hasProperty('AppCenterAppSecret') ? '"'+project.property('AppCenterAppSecret')+'"': '"Define AppCenter App Secret"'
-def STORAGE_ACCOUNT_CONNECTION_STRING = project.hasProperty('StorageAccountConnectionString') ? '"'+project.property('StorageAccountConnectionString')+'"': '"Define Storage Account Connection String"'
+def IOTHUB_E2E_X509_PRIVATE_KEY_BASE64 = project.hasProperty('IotHubPrivateKeyBase64') ? '"'+project.property('IotHubPrivateKeyBase64')+'"' : '""'
+def IOTHUB_E2E_X509_CERT_BASE64 = project.hasProperty('IotHubPublicCertBase64') ? '"'+project.property('IotHubPublicCertBase64')+'"': '""'
+def IOTHUB_CONNECTION_STRING_ENV_VAR_NAME = project.hasProperty('IotHubConnectionString') ? '"'+project.property('IotHubConnectionString')+'"' : '""'
+def IOTHUB_E2E_X509_THUMBPRINT = project.hasProperty('IotHubThumbprint') ? '"'+project.property('IotHubThumbprint')+'"': '""'
+def IOTHUB_CONN_STRING_INVALIDCERT = project.hasProperty('IotHubInvalidCertConnectionString') ? '"'+project.property('IotHubInvalidCertConnectionString')+'"': '""'
+def APPCENTER_APP_SECRET = project.hasProperty('AppCenterAppSecret') ? '"'+project.property('AppCenterAppSecret')+'"': '""'
+def STORAGE_ACCOUNT_CONNECTION_STRING = project.hasProperty('StorageAccountConnectionString') ? '"'+project.property('StorageAccountConnectionString')+'"': '""'
+def IOT_DPS_CONNECTION_STRING = project.hasProperty('DeviceProvisioningServiceConnectionString') ? '"'+project.property('DeviceProvisioningServiceConnectionString')+'"': '""'
+def DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT = project.hasProperty('DeviceProvisioningServiceGlobalEndpoint') ? '"'+project.property('DeviceProvisioningServiceGlobalEndpoint')+'"': '""'
+def INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT = project.hasProperty('InvalidDeviceProvisioningServiceGlobalEndpoint') ? '"'+project.property('InvalidDeviceProvisioningServiceGlobalEndpoint')+'"': '""'
+def INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING = project.hasProperty('InvalidDeviceProvisioningServiceConnectionString') ? '"'+project.property('InvalidDeviceProvisioningServiceConnectionString')+'"': '""'
+def DEVICE_PROVISIONING_SERVICE_ID_SCOPE = project.hasProperty('DeviceProvisioningServiceIdScope') ? '"'+project.property('DeviceProvisioningServiceIdScope')+'"': '""'
 
 def STRING='String'
 //***********************************************************************************************//
@@ -33,6 +38,11 @@ android {
             buildConfigField STRING, 'IotHubInvalidCertConnectionString', IOTHUB_CONN_STRING_INVALIDCERT
             buildConfigField STRING, 'AppCenterAppSecret', APPCENTER_APP_SECRET
             buildConfigField STRING, 'StorageAccountConnectionString', STORAGE_ACCOUNT_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceConnectionString', IOT_DPS_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceGlobalEndpoint', DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT
+            buildConfigField STRING, 'InvalidDeviceProvisioningServiceGlobalEndpoint', INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT
+            buildConfigField STRING, 'InvalidDeviceProvisioningServiceConnectionString', INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceIdScope', DEVICE_PROVISIONING_SERVICE_ID_SCOPE
         }
     }
     buildTypes {
