+++ iot-e2e-tests/android/things/build.gradle
@@ -7,6 +7,11 @@ def IOTHUB_CONNECTION_STRING_ENV_VAR_NAME = project.hasProperty('IotHubConnectio
 def IOTHUB_E2E_X509_THUMBPRINT = project.hasProperty('IotHubThumbprint') ? '"'+project.property('IotHubThumbprint')+'"': '"Define IotHub Cert Thumbprint"'
 def IOTHUB_CONN_STRING_INVALIDCERT = project.hasProperty('IotHubInvalidCertConnectionString') ? '"'+project.property('IotHubInvalidCertConnectionString')+'"': '"Define IotHub Invalid Cert Connection"'
 def STORAGE_ACCOUNT_CONNECTION_STRING = project.hasProperty('StorageAccountConnectionString') ? '"'+project.property('StorageAccountConnectionString')+'"': '"Define Storage Account Connection String"'
+def IOT_DPS_CONNECTION_STRING = project.hasProperty('DeviceProvisioningServiceConnectionString') ? '"'+project.property('DeviceProvisioningServiceConnectionString')+'"': '"Define DeviceProvisioningServiceConnectionString"'
+def DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT = project.hasProperty('DeviceProvisioningServiceGlobalEndpoint') ? '"'+project.property('DeviceProvisioningServiceGlobalEndpoint')+'"': '"Define DeviceProvisioningServiceGlobalEndpoint"'
+def INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT = project.hasProperty('InvalidDeviceProvisioningServiceGlobalEndpoint') ? '"'+project.property('InvalidDeviceProvisioningServiceGlobalEndpoint')+'"': '"Define InvalidDeviceProvisioningServiceGlobalEndpoint"'
+def INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING = project.hasProperty('InvalidDeviceProvisioningServiceConnectionString') ? '"'+project.property('InvalidDeviceProvisioningServiceConnectionString')+'"': '"Define InvalidDeviceProvisioningServiceConnectionString"'
+def DEVICE_PROVISIONING_SERVICE_ID_SCOPE = project.hasProperty('DeviceProvisioningServiceIdScope') ? '"'+project.property('DeviceProvisioningServiceIdScope')+'"': '"Define DeviceProvisioningServiceIdScope"'
 
 def STRING='String'
 //***********************************************************************************************//
@@ -31,6 +36,11 @@ android {
             buildConfigField STRING, 'IotHubThumbprint', IOTHUB_E2E_X509_THUMBPRINT
             buildConfigField STRING, 'IotHubInvalidCertConnectionString', IOTHUB_CONN_STRING_INVALIDCERT
             buildConfigField STRING, 'StorageAccountConnectionString', STORAGE_ACCOUNT_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceConnectionString', IOT_DPS_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceGlobalEndpoint', DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT
+            buildConfigField STRING, 'InvalidDeviceProvisioningServiceGlobalEndpoint', INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT
+            buildConfigField STRING, 'InvalidDeviceProvisioningServiceConnectionString', INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING
+            buildConfigField STRING, 'DeviceProvisioningServiceIdScope', DEVICE_PROVISIONING_SERVICE_ID_SCOPE
         }
     }
