+++ iot-e2e-tests/android/things/src/androidTest/java/com/microsoft/azure/sdk/iot/androidthings/provisioning/ProvisioningClientThingsRunner.java
@@ -1,0 +1,47 @@
+/*
+ *  Copyright (c) Microsoft. All rights reserved.
+ *  Licensed under the MIT license. See LICENSE file in the project root for full license information.
+ */
+
+package com.microsoft.azure.sdk.iot.androidthings.provisioning;
+
+import com.microsoft.azure.sdk.iot.androidthings.BuildConfig;
+import com.microsoft.azure.sdk.iot.androidthings.helper.TestGroupA;
+import com.microsoft.azure.sdk.iot.common.helpers.Rerun;
+import com.microsoft.azure.sdk.iot.common.setup.ProvisioningCommon;
+import com.microsoft.azure.sdk.iot.common.tests.provisioning.ProvisioningTests;
+import com.microsoft.azure.sdk.iot.provisioning.device.ProvisioningDeviceClientTransportProtocol;
+
+import org.junit.Rule;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Collection;
+
+@TestGroupA
+@RunWith(Parameterized.class)
+public class ProvisioningClientThingsRunner extends ProvisioningTests
+{
+    @Rule
+    public Rerun count = new Rerun(3);
+
+    public ProvisioningClientThingsRunner(ProvisioningDeviceClientTransportProtocol protocol, AttestationType attestationType)
+    {
+        super(protocol, attestationType);
+    }
+
+    //This function is run before even the @BeforeClass annotation, so it is used as the @BeforeClass method
+    @Parameterized.Parameters(name = "{0} with {1}")
+    public static Collection inputs() throws Exception
+    {
+        iotHubConnectionString = BuildConfig.IotHubConnectionString;
+        provisioningServiceConnectionString = BuildConfig.DeviceProvisioningServiceConnectionString;
+        provisioningServiceGlobalEndpoint = BuildConfig.DeviceProvisioningServiceGlobalEndpoint;
+        provisioningServiceIdScope = BuildConfig.DeviceProvisioningServiceIdScope;
+        tpmSimulatorIpAddress = ""; //TPM tests lack infrastructure to be used currently, could add later
+        provisioningServiceGlobalEndpointWithInvalidCert = BuildConfig.InvalidDeviceProvisioningServiceGlobalEndpoint;
+        provisioningServiceWithInvalidCertConnectionString = BuildConfig.InvalidDeviceProvisioningServiceConnectionString;
+
+        return ProvisioningCommon.inputs();
+    }
+}
