+++ jenkins/android_java.cmd
@@ -18,5 +18,5 @@ if errorlevel 1 goto :eof
 cd %build-root%\iot-e2e-tests\android
 call gradle wrapper
 call gradlew :clean :app:clean :app:assembleDebug :things:clean :things:assembleDebug
-call gradlew :app:assembleDebugAndroidTest -PIotHubConnectionString=%IOTHUB_CONNECTION_STRING% -PIotHubPublicCertBase64=%IOTHUB_E2E_X509_CERT_BASE64% -PIotHubPrivateKeyBase64=%IOTHUB_E2E_X509_PRIVATE_KEY_BASE64% -PIotHubThumbprint=%IOTHUB_E2E_X509_THUMBPRINT% -PIotHubInvalidCertConnectionString=%IOTHUB_CONN_STRING_INVALIDCERT% -PAppCenterAppSecret=%APPCENTER_APP_SECRET%
-call gradlew :things:assembleDebugAndroidTest -PIotHubConnectionString=%IOTHUB_CONNECTION_STRING% -PIotHubPublicCertBase64=%IOTHUB_E2E_X509_CERT_BASE64% -PIotHubPrivateKeyBase64=%IOTHUB_E2E_X509_PRIVATE_KEY_BASE64% -PIotHubThumbprint=%IOTHUB_E2E_X509_THUMBPRINT% -PIotHubInvalidCertConnectionString=%IOTHUB_CONN_STRING_INVALIDCERT% -PAppCenterAppSecret=%APPCENTER_APP_SECRET%
+call gradlew :app:assembleDebugAndroidTest -PIotHubConnectionString=%IOTHUB_CONNECTION_STRING% -PIotHubPublicCertBase64=%IOTHUB_E2E_X509_CERT_BASE64% -PIotHubPrivateKeyBase64=%IOTHUB_E2E_X509_PRIVATE_KEY_BASE64% -PIotHubThumbprint=%IOTHUB_E2E_X509_THUMBPRINT% -PIotHubInvalidCertConnectionString=%IOTHUB_CONN_STRING_INVALIDCERT% -PAppCenterAppSecret=%APPCENTER_APP_SECRET% -PDeviceProvisioningServiceConnectionString=%IOT_DPS_CONNECTION_STRING% -PDeviceProvisioningServiceIdScope=%DEVICE_PROVISIONING_SERVICE_ID_SCOPE% -PDeviceProvisioningServiceGlobalEndpoint=%DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT% -PInvalidDeviceProvisioningServiceGlobalEndpoint=%INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT% -PInvalidDeviceProvisioningServiceConnectionString=%INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING%
+call gradlew :things:assembleDebugAndroidTest -PIotHubConnectionString=%IOTHUB_CONNECTION_STRING% -PIotHubPublicCertBase64=%IOTHUB_E2E_X509_CERT_BASE64% -PIotHubPrivateKeyBase64=%IOTHUB_E2E_X509_PRIVATE_KEY_BASE64% -PIotHubThumbprint=%IOTHUB_E2E_X509_THUMBPRINT% -PIotHubInvalidCertConnectionString=%IOTHUB_CONN_STRING_INVALIDCERT% -PAppCenterAppSecret=%APPCENTER_APP_SECRET% -PDeviceProvisioningServiceConnectionString=%IOT_DPS_CONNECTION_STRING% -PDeviceProvisioningServiceIdScope=%DEVICE_PROVISIONING_SERVICE_ID_SCOPE% -PDeviceProvisioningServiceGlobalEndpoint=%DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT% -PInvalidDeviceProvisioningServiceGlobalEndpoint=%INVALID_DEVICE_PROVISIONING_SERVICE_GLOBAL_ENDPOINT% -PInvalidDeviceProvisioningServiceConnectionString=%INVALID_DEVICE_PROVISIONING_SERVICE_CONNECTION_STRING%
